Pass --create argument to pg_dump to drop/recreate the DB at the beginning of the dump
This should solve an issue where if you create a new table Foo on a
branch, then switch back to master, Foo continues to exists as the
"--clean" argument didn't know anything about Foo when master was last
dumped, so there's no command to drop the table. Foo thus live on in the
master DB, even though it was new to the feature branch.